% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bIbWNoRm.R
\name{bIbWNoRm}
\alias{bIbWNoRm}
\title{Within and Between (bIbW) Samples Normalization of RNA-seq data}
\usage{
bIbWNoRm(
  Data,
  Clusters = NULL,
  MaxClusterSize = 3000,
  scaling = NULL,
  filter = FALSE,
  FilterExpression = 0,
  FilterCellNum = 10,
  numGeneforEst = 2000,
  divideforFast = TRUE,
  numDivide = NULL,
  Method = c("KernSmooth", "SJ", "ucv", "bcv", "nrd"),
  cutoff = 1.5,
  BPPARAM = SerialParam()
)
}
\arguments{
\item{Data}{as a Matrix with rows as genes (features) and column as cells (samples).}

\item{Conditions}{and Methods for variance selection (Bandwidth).}
}
\value{
Output Normalized data and scale factor.
}
\description{
Within and Between (bIbW) Samples Normalization of RNA-seq data
}
\examples{
library(scran)
library(qsmooth)
library(Qtools)
library(dynutils)
library(BiocParallel)
set.seed(12345)
G <- 2000; n <- 600 # G: number of genes, n: number of cells
mu <- rgamma(G, shape = 2, rate = 2)
NB_cell <- function(j) rnbinom(G, size = 0.1, mu = mu)
SimulatedData <- sapply(1:n, NB_cell)
colnames(SimulatedData) <- paste("c", 1:n, sep = "_")
rownames(SimulatedData) <- paste("g", 1:G, sep = "_")
Conditions = rep(c(1,2), each= 300)
Clusters <- quickCluster(SimulatedData)
Result <- bIbWNoRm(Data = SimulatedData, Clusters = Clusters, Method= "KernSmooth")
}
